move stoppers to separate class
